Changed streamID generating as a incremental short value to System.currentmillis() Long value. This will resolve having duplicate streamId from the receiver and sender.

+++ android/android-libraries/android-zirk-proxy/src/main/java/com/bezirk/middleware/android/FileStreamController.java
@@ -11,10 +11,10 @@ import com.bezirk.middleware.streaming.StreamController;
 class FileStreamController extends StreamController {
 
     //primary key for streaming.
-    private final Short streamId;
+    private final Long streamId;
 
     //constructor
-    FileStreamController(Short streamId){
+    FileStreamController(Long streamId){
         this.streamId = streamId;
     }
 
@@ -28,7 +28,7 @@ class FileStreamController extends StreamController {
         */
     }
 
-    public Short getStreamId() {
+    public Long getStreamId() {
         return streamId;
     }
